# backend/openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "RainForge API",
        "description": "Complete Rooftop Rainwater Harvesting Platform API\n\n## Features\n- **Assessment**: Instant RWH assessment with 3 scenarios\n- **PDF Generation**: Engineer-signed reports with QR verification\n- **Auction**: Reverse auction for installer bidding\n- **Allocation**: Smart installer allocation (RPI-weighted)\n- **Escrow**: Milestone-based payment release\n- **Verification**: Fraud-resistant geo-verified photo uploads\n- **Monitoring**: Real-time IoT telemetry\n- **Public Dashboard**: Ward-level transparency reporting",
        "version": "1.0.0",
        "contact": {
            "name": "RainForge Team",
            "email": "support@rainforge.in"
        }
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local development"
        }
    ],
...
    "paths": {
        "/assess": {
            "post": {
                "tags": [
                    "Assessment"
                ],
                "summary": "Create RWH Assessment",
                "description": "Create a complete rainwater harvesting assessment with 3 scenarios: cost_optimized, max_capture, and dry_season.",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssessmentRequest"
                            },
                            "example": {
                                "site_id": "SITE001",
                                "address": "Municipal School Sector 5, Goa",
                                "lat": 15.0,
                                "lng": 73.0,
                                "roof_area_sqm": 120,
                                "roof_material": "concrete",
                                "demand_l_per_day": 200,
                                "state": "Goa",
                                "city": "Panaji"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assessment created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssessmentResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/assess/{assessment_id}": {
            "get": {
                "tags": [
                    "Assessment"
                ],
                "summary": "Get Assessment Details",
                "parameters": [
                    {
                        "name": "assessment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Assessment details"
                    },
                    "404": {
                        "description": "Assessment not found"
                    }
                }
            }
        },
        "/assess/{assessment_id}/pdf": {
            "get": {
                "tags": [
                    "Assessment"
                ],
                "summary": "Download Assessment PDF",
                "description": "Generate and download PDF with QR verification code and engineer signature block.",
                "parameters": [
                    {
                        "name": "assessment_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PDF file",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "Assessment not found"
                    }
                }
            }
        },
...
        "/verify/{verification_code}": {
            "get": {
                "tags": [
                    "Verification"
                ],
                "summary": "Verify Assessment (QR)",
                "description": "Verify an assessment using QR code. Returns assessment details and audit trail.",
                "parameters": [
                    {
                        "name": "verification_code",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verified assessment with audit trail"
                    },
                    "404": {
                        "description": "Assessment not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AssessmentRequest": {
                "type": "object",
                "required": [
                    "site_id",
                    "address",
                    "lat",
                    "lng",
                    "roof_area_sqm"
                ],
                "properties": {
                    "site_id": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "minimum": -90,
                        "maximum": 90
                    },
                    "lng": {
                        "type": "number",
                        "minimum": -180,
                        "maximum": 180
                    },
                    "roof_area_sqm": {
                        "type": "number",
                        "minimum": 1
                    },
                    "roof_material": {
                        "type": "string",
                        "enum": [
                            "concrete",
                            "metal",
                            "tiles",
                            "thatched",
                            "asbestos"
                        ]
                    },
                    "demand_l_per_day": {
                        "type": "number"
                    },
                    "floors": {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 20
                    },
                    "people": {
                        "type": "integer",
                        "minimum": 1
                    },
                    "state": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    }
                }
            },
            "Scenario": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "tank_liters": {
                        "type": "integer"
                    },
                    "cost_inr": {
                        "type": "number"
                    },
                    "capture_liters": {
                        "type": "number"
                    },
                    "coverage_days": {
                        "type": "integer"
                    },
                    "roi_years": {
                        "type": "number"
                    }
                }
            },
            "AssessmentResponse": {
                "type": "object",
                "properties": {
                    "assessment_id": {
                        "type": "string"
                    },
                    "site_id": {
                        "type": "string"
                    },
                    "annual_rainfall_mm": {
                        "type": "number"
                    },
                    "annual_yield_liters": {
                        "type": "number"
                    },
                    "scenarios": {
                        "type": "object",
                        "properties": {
                            "cost_optimized": {
                                "$ref": "#/components/schemas/Scenario"
                            },
                            "max_capture": {
                                "$ref": "#/components/schemas/Scenario"
                            },
                            "dry_season": {
                                "$ref": "#/components/schemas/Scenario"
                            }
                        }
                    },
                    "recommended_scenario": {
                        "type": "string"
                    },
                    "subsidy_pct": {
                        "type": "number"
                    },
                    "subsidy_amount_inr": {
                        "type": "number"
                    },
                    "co2_avoided_kg": {
                        "type": "number"
                    },
                    "pdf_url": {
                        "type": "string"
                    },
                    "verify_url": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            },
...
        }
    }
}